requirements.txt
pandas
numpy
PILasOPENCV
streamlit
httpx-oauth